Added database package for future change, tiles rotation fixed, added tests

+++ src/main/java/com/agawrysiuk/huntbeginsspringboot/model/FloorTile.java
@@ -4,15 +4,13 @@ public interface FloorTile {
 
     int getId();
 
-    //set up exits from file?
     void setExits(int... exits);
 
-    //exits necessary or not?
     int[] getExits();
 
     String getName();
 
     void rotate();
 
-    int getRotate();
+    double getRotate();
 }
